requirements.txt
cffi==1.6.0
cryptography==1.3.2
Django==1.9.6
djangorestframework==3.3.3
enum34==1.1.6
idna==2.1
ipaddress==1.0.16
ndg-httpsclient==0.4.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
six==1.10.0
